refactor: restructure chat component and introduce ExamplePrompts for better UI
