+++ app/components/chat/BaseChat.tsx
@@ -14,23 +14,13 @@ import { Messages } from './Messages.client';
 import { SendButton } from './SendButton.client';
 import { APIKeyManager } from './APIKeyManager';
 import Cookies from 'js-cookie';
-import { toast } from 'react-toastify';
 import * as Tooltip from '@radix-ui/react-tooltip';
 
 import styles from './BaseChat.module.scss';
 import type { ProviderInfo } from '~/utils/types';
-import { ExportChatButton } from '~/components/chat/ExportChatButton';
-
-const EXAMPLE_PROMPTS = [
-  { text: 'Build a todo app in React using Tailwind' },
-  { text: 'Build a simple blog using Astro' },
-  { text: 'Create a cookie consent form using Material UI' },
-  { text: 'Make a space invaders game' },
-  { text: 'How do I center a div?' },
-];
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const providerList = PROVIDER_LIST;
+import { ExportChatButton } from '~/components/chat/chatExportAndImport/ExportChatButton';
+import { ImportButtons } from '~/components/chat/chatExportAndImport/ImportButtons';
+import { ExamplePrompts } from '~/components/chat/ExamplePrompts';
 
 // @ts-ignore TODO: Introduce proper types
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -126,7 +116,6 @@ export const BaseChat = React.forwardRef<HTMLDivElement, BaseChatProps>(
     const TEXTAREA_MAX_HEIGHT = chatStarted ? 400 : 200;
     const [apiKeys, setApiKeys] = useState<Record<string, string>>({});
     const [modelList, setModelList] = useState(MODEL_LIST);
-    const [isVisible, setIsVisible] = useState(true);
 
     useEffect(() => {
       // Load API keys from cookies on component mount
@@ -169,84 +158,22 @@ export const BaseChat = React.forwardRef<HTMLDivElement, BaseChatProps>(
       }
     };
 
-    const chatImportButton = !chatStarted && (
-      <div className="flex flex-col items-center justify-center flex-1 p-4">
-        <input
-          type="file"
-          id="chat-import"
-          className="hidden"
-          accept=".json"
-          onChange={async (e) => {
-            const file = e.target.files?.[0];
-
-            if (file && importChat) {
-              try {
-                const reader = new FileReader();
-
-                reader.onload = async (e) => {
-                  try {
-                    const content = e.target?.result as string;
-                    const data = JSON.parse(content);
-
-                    if (!Array.isArray(data.messages)) {
-                      toast.error('Invalid chat file format');
-                    }
-
-                    await importChat(data.description, data.messages);
-                    toast.success('Chat imported successfully');
-                  } catch (error: unknown) {
-                    if (error instanceof Error) {
-                      toast.error('Failed to parse chat file: ' + error.message);
-                    } else {
-                      toast.error('Failed to parse chat file');
-                    }
-                  }
-                };
-                reader.onerror = () => toast.error('Failed to read chat file');
-                reader.readAsText(file);
-              } catch (error) {
-                toast.error(error instanceof Error ? error.message : 'Failed to import chat');
-              }
-              e.target.value = ''; // Reset file input
-            } else {
-              toast.error('Something went wrong');
-            }
-          }}
-        />
-        <div className="flex flex-col items-center gap-4 max-w-2xl text-center">
-          <div className="flex gap-2">
-            <button
-              onClick={() => {
-                const input = document.getElementById('chat-import');
-                input?.click();
-              }}
-              className="px-4 py-2 rounded-lg border border-bolt-elements-borderColor bg-bolt-elements-prompt-background text-bolt-elements-textPrimary hover:bg-bolt-elements-background-depth-3 transition-all flex items-center gap-2"
-            >
-              <div className="i-ph:upload-simple" />
-              Import Chat
-            </button>
-          </div>
-        </div>
-      </div>
-    );
-
     const baseChat = (
       <div
         ref={ref}
         className={classNames(
           styles.BaseChat,
-          'relative flex flex-col lg:flex-row h-full w-full overflow-hidden bg-bolt-elements-background-depth-1 relative',
+          'relative flex flex-col lg:flex-row h-full w-full overflow-hidden bg-bolt-elements-background-depth-1',
         )}
         data-chat-visible={showChat}
       >
-        <div className="h-10 w-3/5 -top-10 rounded-full bg-blue-500/50 dark:bg-blue-600/40 absolute blur-3xl"></div>
         <ClientOnly>{() => <Menu />}</ClientOnly>
         <div ref={scrollRef} className="flex flex-col lg:flex-row overflow-y-auto w-full h-full">
           <div className={classNames(styles.Chat, 'flex flex-col flex-grow lg:min-w-[var(--chat-min-width)] h-full')}>
             {!chatStarted && (
-              <div id="intro" className="mt-[26vh] max-w-5xl mx-auto text-center px-4 lg:px-0">
-                <h1 className="text-3xl lg:text-6xl font-bold bg-gradient-to-r to-bolt-elements-textSecondary via-bolt-elements-textPrimary bg-clip-text text-transparent py-2 from-bolt-elements-textPrimary mb-4 animate-fade-in tracking-tight">
-                  Where ideas begin ✨
+              <div id="intro" className="mt-[26vh] max-w-chat mx-auto text-center px-4 lg:px-0">
+                <h1 className="text-3xl lg:text-6xl font-bold text-bolt-elements-textPrimary mb-4 animate-fade-in">
+                  Where ideas begin
                 </h1>
                 <p className="text-md lg:text-xl mb-8 text-bolt-elements-textSecondary animate-fade-in animation-delay-200">
                   Bring ideas to life in seconds or get help on existing projects.
@@ -272,48 +199,30 @@ export const BaseChat = React.forwardRef<HTMLDivElement, BaseChatProps>(
               </ClientOnly>
               <div
                 className={classNames(
-                  'dark:bg-neutral-900/90 bg-neutral-50/85 backdrop-blur-md p-3 rounded-lg border border-bolt-elements-borderColor relative w-full max-w-chat mx-auto z-prompt mb-6 relative',
+                  ' bg-bolt-elements-background-depth-2 p-3 rounded-lg border border-bolt-elements-borderColor relative w-full max-w-chat mx-auto z-prompt mb-6',
                   {
                     'sticky bottom-2': chatStarted,
                   },
                 )}
               >
-                {chatStarted && (
-                  <button
-                    onClick={() => setIsVisible(!isVisible)}
-                    className="dark:text-white text-black dark:bg-neutral-900/50 backdrop-blur p-1 shadow rounded mb-2 absolute -top-6"
-                  >
-                    <div
-                      className={classNames('i-ph:caret-line-down size-4', {
-                        'transform rotate-180': !isVisible,
-                      })}
-                    ></div>
-                  </button>
-                )}
-                <div
-                  className={classNames({
-                    hidden: !isVisible,
-                  })}
-                >
-                  <ModelSelector
-                    key={provider?.name + ':' + modelList.length}
-                    model={model}
-                    setModel={setModel}
-                    modelList={modelList}
-                    provider={provider}
-                    setProvider={setProvider}
-                    providerList={PROVIDER_LIST}
-                    apiKeys={apiKeys}
-                  />
+                <ModelSelector
+                  key={provider?.name + ':' + modelList.length}
+                  model={model}
+                  setModel={setModel}
+                  modelList={modelList}
+                  provider={provider}
+                  setProvider={setProvider}
+                  providerList={PROVIDER_LIST}
+                  apiKeys={apiKeys}
+                />
 
-                  {provider && (
-                    <APIKeyManager
-                      provider={provider}
-                      apiKey={apiKeys[provider.name] || ''}
-                      setApiKey={(key) => updateApiKey(provider.name, key)}
-                    />
-                  )}
-                </div>
+                {provider && (
+                  <APIKeyManager
+                    provider={provider}
+                    apiKey={apiKeys[provider.name] || ''}
+                    setApiKey={(key) => updateApiKey(provider.name, key)}
+                  />
+                )}
 
                 <div
                   className={classNames(
@@ -398,27 +307,8 @@ export const BaseChat = React.forwardRef<HTMLDivElement, BaseChatProps>(
                 </div>
               </div>
             </div>
-            {chatImportButton}
-            {!chatStarted && (
-              <div id="examples" className="relative w-full max-w-2xl mx-auto mt-6 flex justify-center">
-                <div className="flex flex-wrap items-center justify-center gap-2 [mask-image:linear-gradient(to_bottom,black_0%,transparent_400%)] hover:[mask-image:none] w-full border mb-12">
-                  {EXAMPLE_PROMPTS.map((examplePrompt, index) => {
-                    return (
-                      <button
-                        key={index}
-                        onClick={(event) => {
-                          sendMessage?.(event, examplePrompt.text);
-                        }}
-                        className="group flex items-center bg-zinc-100 dark:bg-gray-900 px-3 rounded-full  gap-2 justify-center text-bolt-elements-textTertiary hover:text-bolt-elements-textPrimary transition-theme text-sm"
-                      >
-                        {examplePrompt.text}
-                        <div className="i-ph:arrow-bend-down-left" />
-                      </button>
-                    );
-                  })}
-                </div>
-              </div>
-            )}
+            {!chatStarted && ImportButtons(importChat)}
+            {!chatStarted && ExamplePrompts(sendMessage)}
           </div>
           <ClientOnly>{() => <Workbench chatStarted={chatStarted} isStreaming={isStreaming} />}</ClientOnly>
         </div>
